Fix issues mentioned in cr
