--- components/director/examples/add-document/add-document.graphql
@@ -1,63 +1,0 @@
-# Code generated by Compass integration tests, DO NOT EDIT.
-mutation {
-  result: addDocument(
-    applicationID: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
-    in: {
-      title: "new-document"
-      displayName: "new-document-display-name"
-      description: "new-description"
-      format: MARKDOWN
-    }
-  ) {
-    id
-    applicationID
-    title
-    displayName
-    description
-    format
-    kind
-    data
-    fetchRequest {
-      url
-      auth {
-        credential {
-          ... on BasicCredentialData {
-            username
-            password
-          }
-          ... on OAuthCredentialData {
-            clientId
-            clientSecret
-            url
-          }
-        }
-        additionalHeaders
-        additionalQueryParams
-        requestAuth {
-          csrf {
-            tokenEndpointURL
-            credential {
-              ... on BasicCredentialData {
-                username
-                password
-              }
-              ... on OAuthCredentialData {
-                clientId
-                clientSecret
-                url
-              }
-            }
-            additionalHeaders
-            additionalQueryParams
-          }
-        }
-      }
-      mode
-      filter
-      status {
-        condition
-        timestamp
-      }
-    }
-  }
-}
