--- components/director/examples/add-event-definition/add-event-definition.graphql
@@ -1,69 +1,0 @@
-# Code generated by Compass integration tests, DO NOT EDIT.
-mutation {
-  result: addEventDefinition(
-    applicationID: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
-    in: {
-      name: "new-event-api"
-      spec: { type: ASYNC_API, fetchRequest: { url: "foo.bar" }, format: YAML }
-    }
-  ) {
-    id
-    applicationID
-    name
-    description
-    group
-    spec {
-      data
-      type
-      format
-      fetchRequest {
-        url
-        auth {
-          credential {
-            ... on BasicCredentialData {
-              username
-              password
-            }
-            ... on OAuthCredentialData {
-              clientId
-              clientSecret
-              url
-            }
-          }
-          additionalHeaders
-          additionalQueryParams
-          requestAuth {
-            csrf {
-              tokenEndpointURL
-              credential {
-                ... on BasicCredentialData {
-                  username
-                  password
-                }
-                ... on OAuthCredentialData {
-                  clientId
-                  clientSecret
-                  url
-                }
-              }
-              additionalHeaders
-              additionalQueryParams
-            }
-          }
-        }
-        mode
-        filter
-        status {
-          condition
-          timestamp
-        }
-      }
-    }
-    version {
-      value
-      deprecated
-      deprecatedSince
-      forRemoval
-    }
-  }
-}
