+++ components/director/examples/query-api-definition/query-api-definition.graphql
@@ -1,17 +1,63 @@
 # Code generated by Compass integration tests, DO NOT EDIT.
 query {
   result: application(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
-    apiDefinition(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
-      id
-      name
-      applicationID
-      description
-      spec {
-        data
-        format
-        type
-        fetchRequest {
-          url
+    package(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+      apiDefinition(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+        id
+        name
+        applicationID
+        description
+        spec {
+          data
+          format
+          type
+          fetchRequest {
+            url
+            auth {
+              credential {
+                ... on BasicCredentialData {
+                  username
+                  password
+                }
+                ... on OAuthCredentialData {
+                  clientId
+                  clientSecret
+                  url
+                }
+              }
+              additionalHeaders
+              additionalQueryParams
+              requestAuth {
+                csrf {
+                  tokenEndpointURL
+                  credential {
+                    ... on BasicCredentialData {
+                      username
+                      password
+                    }
+                    ... on OAuthCredentialData {
+                      clientId
+                      clientSecret
+                      url
+                    }
+                  }
+                  additionalHeaders
+                  additionalQueryParams
+                }
+              }
+            }
+            mode
+            filter
+            status {
+              condition
+              timestamp
+            }
+          }
+        }
+        targetURL
+        group
+        auths {
+          runtimeID
           auth {
             credential {
               ... on BasicCredentialData {
@@ -45,19 +91,8 @@ query {
               }
             }
           }
-          mode
-          filter
-          status {
-            condition
-            timestamp
-          }
         }
-      }
-      targetURL
-      group
-      auths {
-        runtimeID
-        auth {
+        defaultAuth {
           credential {
             ... on BasicCredentialData {
               username
@@ -90,45 +125,12 @@ query {
             }
           }
         }
-      }
-      defaultAuth {
-        credential {
-          ... on BasicCredentialData {
-            username
-            password
-          }
-          ... on OAuthCredentialData {
-            clientId
-            clientSecret
-            url
-          }
+        version {
+          value
+          deprecated
+          deprecatedSince
+          forRemoval
         }
-        additionalHeaders
-        additionalQueryParams
-        requestAuth {
-          csrf {
-            tokenEndpointURL
-            credential {
-              ... on BasicCredentialData {
-                username
-                password
-              }
-              ... on OAuthCredentialData {
-                clientId
-                clientSecret
-                url
-              }
-            }
-            additionalHeaders
-            additionalQueryParams
-          }
-        }
-      }
-      version {
-        value
-        deprecated
-        deprecatedSince
-        forRemoval
       }
     }
   }
