+++ components/director/examples/query-document/query-document.graphql
@@ -1,0 +1,59 @@
+# Code generated by Compass integration tests, DO NOT EDIT.
+query {
+  result: application(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+    package(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+      document(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+        id
+        applicationID
+        title
+        displayName
+        description
+        format
+        kind
+        data
+        fetchRequest {
+          url
+          auth {
+            credential {
+              ... on BasicCredentialData {
+                username
+                password
+              }
+              ... on OAuthCredentialData {
+                clientId
+                clientSecret
+                url
+              }
+            }
+            additionalHeaders
+            additionalQueryParams
+            requestAuth {
+              csrf {
+                tokenEndpointURL
+                credential {
+                  ... on BasicCredentialData {
+                    username
+                    password
+                  }
+                  ... on OAuthCredentialData {
+                    clientId
+                    clientSecret
+                    url
+                  }
+                }
+                additionalHeaders
+                additionalQueryParams
+              }
+            }
+          }
+          mode
+          filter
+          status {
+            condition
+            timestamp
+          }
+        }
+      }
+    }
+  }
+}
