+++ components/director/examples/query-event-definition/query-event-definition.graphql
@@ -1,64 +1,66 @@
 # Code generated by Compass integration tests, DO NOT EDIT.
 query {
   result: application(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
-    eventDefinition(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
-      id
-      applicationID
-      name
-      description
-      group
-      spec {
-        data
-        type
-        format
-        fetchRequest {
-          url
-          auth {
-            credential {
-              ... on BasicCredentialData {
-                username
-                password
-              }
-              ... on OAuthCredentialData {
-                clientId
-                clientSecret
-                url
-              }
-            }
-            additionalHeaders
-            additionalQueryParams
-            requestAuth {
-              csrf {
-                tokenEndpointURL
-                credential {
-                  ... on BasicCredentialData {
-                    username
-                    password
-                  }
-                  ... on OAuthCredentialData {
-                    clientId
-                    clientSecret
-                    url
-                  }
+    package(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+      eventDefinition(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+        id
+        applicationID
+        name
+        description
+        group
+        spec {
+          data
+          type
+          format
+          fetchRequest {
+            url
+            auth {
+              credential {
+                ... on BasicCredentialData {
+                  username
+                  password
                 }
-                additionalHeaders
-                additionalQueryParams
+                ... on OAuthCredentialData {
+                  clientId
+                  clientSecret
+                  url
+                }
+              }
+              additionalHeaders
+              additionalQueryParams
+              requestAuth {
+                csrf {
+                  tokenEndpointURL
+                  credential {
+                    ... on BasicCredentialData {
+                      username
+                      password
+                    }
+                    ... on OAuthCredentialData {
+                      clientId
+                      clientSecret
+                      url
+                    }
+                  }
+                  additionalHeaders
+                  additionalQueryParams
+                }
               }
             }
-          }
-          mode
-          filter
-          status {
-            condition
-            timestamp
+            mode
+            filter
+            status {
+              condition
+              timestamp
+            }
           }
         }
-      }
-      version {
-        value
-        deprecated
-        deprecatedSince
-        forRemoval
+        version {
+          value
+          deprecated
+          deprecatedSince
+          forRemoval
+        }
       }
     }
   }
