+++ tests/director/api/fixtures.go
@@ -105,3 +105,43 @@ func fixPackageUpdateInput(name string) graphql.PackageUpdateInput {
 		Name: name,
 	}
 }
+
+func fixAPIDefinitionInputWithName(name string) graphql.APIDefinitionInput {
+	return graphql.APIDefinitionInput{
+		Name:      name,
+		TargetURL: "https://target.url",
+		Spec: &graphql.APISpecInput{
+			Format: graphql.SpecFormatJSON,
+			Type:   graphql.APISpecTypeOpenAPI,
+			FetchRequest: &graphql.FetchRequestInput{
+				URL: "https://foo.bar",
+			},
+		},
+	}
+
+}
+
+func fixEventAPIDefinitionInputWithName(name string) graphql.EventDefinitionInput {
+	data := graphql.CLOB("data")
+	return graphql.EventDefinitionInput{Name: name,
+		Spec: &graphql.EventSpecInput{
+			Data:   &data,
+			Type:   graphql.EventSpecTypeAsyncAPI,
+			Format: graphql.SpecFormatJSON,
+		}}
+}
+
+func fixDocumentInputWithName(name string) graphql.DocumentInput {
+	return graphql.DocumentInput{
+		Title:       name,
+		Description: "Detailed description of project",
+		Format:      graphql.DocumentFormatMarkdown,
+		DisplayName: "display-name",
+		FetchRequest: &graphql.FetchRequestInput{
+			URL:    "kyma-project.io",
+			Mode:   ptr.FetchMode(graphql.FetchModePackage),
+			Filter: ptr.String("/docs/README.md"),
+			Auth:   fixBasicAuth(),
+		},
+	}
+}
