+++ tests/director/api/validation_test.go
@@ -331,34 +331,9 @@ func fixEventAPIDefinitionInput() graphql.EventDefinitionInput {
 		}}
 }
 
-func fixEventAPIDefinitionInputWithName(name string) graphql.EventDefinitionInput {
-	data := graphql.CLOB("data")
-	return graphql.EventDefinitionInput{Name: name,
-		Spec: &graphql.EventSpecInput{
-			Data:   &data,
-			Type:   graphql.EventSpecTypeAsyncAPI,
-			Format: graphql.SpecFormatJSON,
-		}}
-}
-
 func fixAPIDefinitionInput() graphql.APIDefinitionInput {
 	return graphql.APIDefinitionInput{
 		Name:      "new-api-name",
-		TargetURL: "https://target.url",
-		Spec: &graphql.APISpecInput{
-			Format: graphql.SpecFormatJSON,
-			Type:   graphql.APISpecTypeOpenAPI,
-			FetchRequest: &graphql.FetchRequestInput{
-				URL: "https://foo.bar",
-			},
-		},
-	}
-
-}
-
-func fixAPIDefinitionInputWithName(name string) graphql.APIDefinitionInput {
-	return graphql.APIDefinitionInput{
-		Name:      name,
 		TargetURL: "https://target.url",
 		Spec: &graphql.APISpecInput{
 			Format: graphql.SpecFormatJSON,
@@ -444,21 +419,6 @@ func TestRegisterApplicationFromTemplate_Validation(t *testing.T) {
 func fixDocumentInput() graphql.DocumentInput {
 	return graphql.DocumentInput{
 		Title:       "Readme",
-		Description: "Detailed description of project",
-		Format:      graphql.DocumentFormatMarkdown,
-		DisplayName: "display-name",
-		FetchRequest: &graphql.FetchRequestInput{
-			URL:    "kyma-project.io",
-			Mode:   ptr.FetchMode(graphql.FetchModePackage),
-			Filter: ptr.String("/docs/README.md"),
-			Auth:   fixBasicAuth(),
-		},
-	}
-}
-
-func fixDocumentInputWithName(name string) graphql.DocumentInput {
-	return graphql.DocumentInput{
-		Title:       name,
 		Description: "Detailed description of project",
 		Format:      graphql.DocumentFormatMarkdown,
 		DisplayName: "display-name",
