+++ tests/director/pkg/gql/graphqlizer.go
@@ -442,8 +442,6 @@ func (g *Graphqlizer) genericToGQL(obj interface{}, tmpl string) (string, error)
 	fm["CredentialRequestAuthInputToGQL"] = g.CredentialRequestAuthInputToGQL
 	fm["PlaceholderDefinitionInputToGQL"] = g.PlaceholderDefinitionInputToGQL
 	fm["TemplateValueInput"] = g.TemplateValueInputToGQL
-	fm["PackageCreateInputToGQL"] = g.PackageCreateInputToGQL
-	fm["PackageUpdateInputToGQL"] = g.PackageUpdateInputToGQL
 
 	t, err := template.New("tmpl").Funcs(fm).Parse(tmpl)
 	if err != nil {
